Fixed the problems in aStarSearch.
Added initialization for the datastructures.

Reorganized the functions and datastructures, since the compareFValue function wouldn't
otherwise be used inside the heap.

Modified compareFValue to increase performance.

Included a modified version of the Collections.Heap datastructure, to
modify the contains function and to implement a update function.

Added a g-cost value and a predecessor for the start node.

result.cost is now set to the g-cost of the last node in the path, instead
of the currentN which would then be the starting node.

Added the updating of nodes if there already inside of the Heap.

Also included some comments, to make it easier to see what's what.