Modified the Planner
Moved a lot of functions into the classes.
Removed the "unnecessary" variables from the StateNode class and replaced
them with the needed parts.
Added a variable to store the string version of the StateNode to save time.
Merged some of the functions into outgoingEdges to remove extra checks.
Modified some functions to not require a WorldState and to fix some errors.

Added a function to help with determining the actions needed to be
performed after finding the path. (This replaced the move variable inside
of the StateNode class)
Changed the Comments to reflect not needing a world state.

Note that some of the parts being pushed as actions are only for testing purposes.